refactor(agent-core): move pure agent + diff modules out of web app
