+++ packages/agent-core/src/jsonNodes.ts
