+++ packages/agent-core/src/robustDiffStrategy.ts
